admin_staff/requirements copy.txt
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
Babel==2.12.1
beautifulsoup4==4.8.2
billiard==4.1.0
celery==5.3.0
certifi==2019.11.28
cffi==1.15.1
channels==4.0.0
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
cryptography==41.0.1
cssselect2==0.7.0
distro==1.4.0
dj-database-url==2.0.0
Django==4.2.1
django-bootstrap-icons==0.8.3
django-ckeditor==6.5.1
django-ckeditor-5==0.2.8
django-countries==7.5.1
django-htmx==1.14.0
django-jazzmin==2.6.0
django-jazzmin-admin-rangefilter==1.0.0
django-js-asset==2.0.0
django-money==3.1.0
django-ninja==0.22.1
entrypoints==0.3
html5lib==1.1
httplib2==0.14.0
idna==2.8
IMDbPY==6.8
keyring==18.0.1
kombu==5.3.0
lxml==4.9.2
numpy==1.24.3
oauthlib==3.1.0
oscrypto==1.3.0
packaging==20.3
pandas==2.0.2
pexpect==4.6.0
Pillow==9.5.0
prompt-toolkit==3.0.38
protobuf==3.6.1
py-moneyed==2.0
pycparser==2.21
pydantic==1.10.8
Pygments==2.3.1
pyHanko==0.18.1
pyhanko-certvalidator==0.22.0
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pypdf==3.9.1
pypng==0.20220715.0
pyRFC3339==1.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
pyxdg==0.26
qrcode==7.4.2
reportlab==3.6.13
requests==2.31.0
requests-file==1.4.3
requests-unixsocket==0.2.0
six==1.14.0
soupsieve==1.9.5
sqlparse==0.4.4
tldextract==2.2.1
typing_extensions==4.6.3
Unidecode==1.1.1
uritools==4.0.1
urllib3==1.25.8
whitenoise==6.4.0
xhtml2pdf==0.2.11
psycopg2-binary==2.9.6
